+++ src/commands/hardis/project/deploy/sources/dx.ts
@@ -21,6 +21,7 @@ import { getCurrentGitBranch, isCI, uxLog } from "../../../../../common/utils";
 import { getConfig } from "../../../../../config";
 import { forceSourceDeploy } from "../../../../../common/utils/deployUtils";
 import { promptOrg } from "../../../../../common/utils/orgUtils";
+import { getApexTestClasses } from "../../../../../common/utils/classUtils";
 import { restoreListViewMine } from "../../../../../common/utils/orgConfigUtils";
 import { NotifProvider } from "../../../../../common/notifProviderRoot";
 import { GitProvider } from "../../../../../common/gitProvider";
@@ -156,8 +157,12 @@ If you need to increase the deployment waiting time (force:source:deploy --wait 
     testlevel: flags.enum({
       char: "l",
       default: "RunLocalTests",
-      options: ["NoTestRun", "RunSpecifiedTests", "RunLocalTests", "RunAllTestsInOrg"],
-      description: messages.getMessage("testLevel"),
+      options: ["NoTestRun", "RunSpecifiedTests", "RunRepositoryTests", "RunLocalTests", "RunAllTestsInOrg"],
+      description: messages.getMessage("testLevelExtended"),
+    }),
+    runtests: flags.string({
+      char: "r",
+      description: messages.getMessage("runtests"),
     }),
     packagexml: flags.string({
       char: "p",
@@ -190,7 +195,30 @@ If you need to increase the deployment waiting time (force:source:deploy --wait 
   public async run(): Promise<AnyJson> {
     this.configInfo = await getConfig("branch");
     const check = this.flags.check || false;
+
+    const givenTestlevel = this.flags.testlevel || this.configInfo.testLevel || "";
+    let testClasses = this.flags.runtests || this.configInfo.runtests || "";
+
+    // Auto-detect all APEX test classes within project in order to run "dynamic" RunSpecifiedTests deployment
+    if(givenTestlevel === 'RunRepositoryTests') {
+      const testClassList = await getApexTestClasses();
+      if(Array.isArray(testClassList) && testClassList.length) {
+        this.flags.testlevel = 'RunSpecifiedTests';
+        testClasses = testClassList.join();
+      } else {
+        // Default back to RunLocalTests in case if repository has zero tests
+        this.flags.testlevel = 'RunLocalTests'; 
+        testClasses = '';
+      }
+    }
+
     const testlevel = this.flags.testlevel || this.configInfo.testLevel || "RunLocalTests";
+    
+    // Test classes are only valid for RunSpecifiedTests
+    if(testlevel != 'RunSpecifiedTests') {
+      testClasses = '';
+    }
+
     const packageXml = this.flags.packagexml || null;
     this.debugMode = this.flags.debug || false;
 
@@ -231,6 +259,7 @@ If you need to increase the deployment waiting time (force:source:deploy --wait 
     const forceSourceDeployOptions: any = {
       targetUsername: targetUsername,
       conn: this.org?.getConnection(),
+      testClasses: testClasses
     };
     // Get destructiveChanges.xml and add it in options if existing
     const packageDeletedXmlFile =